"refactor: organize barangay dashboard components"
